+++ signal/run.sh
@@ -9,6 +9,9 @@ export SIGNAL_API_URL=http://localhost:8080
 export GIN_MODE=release
 export DEBUG_MODE=$(jq -r 'if .debug_mode == true then "true" else "false" end' /data/options.json)
 export DOWNLOAD_MEDIA=$(jq -r 'if .download_media == true then "true" else "false" end' /data/options.json)
+export HA_NOTIFICATIONS=$(jq -r 'if .ha_notifications == true then "true" else "false" end' /data/options.json)
+export HA_NOTIFICATIONS_PRIVACY=$(jq -r 'if .ha_notifications_privacy == true then "true" else "false" end' /data/options.json)
+export HA_TOKEN=$(jq -r '.ha_token // ""' /data/options.json)
 
 export SIGNAL_CLI_CONFIG_DIR=/data/signal-cli
 mkdir -p /data/signal-cli